\name{getangle}
\alias{getangle}
\title{Find the angle of a plot}
\usage{
getangle(image)
}
\arguments{
  \item{image,}{the image to be worked with}
}
\value{
counterclockwise angle in degrees
}
\description{
'getangle' finds the counterclockwise angle at which the
plot has been rotated.  It asks the user to click on two
ends of the x axis, then two ends of the y axis.
}

