% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idlocator.R
\name{idlocator}
\alias{idlocator}
\title{Digitize a point, or UNDO, or STOP.}
\usage{
idlocator(S = NULL, U = NULL, X = NULL, Y = NULL, bell = TRUE,
  verbose = TRUE, type = "p", col = "red", lwd = par("lwd"), pch = 20,
  cex = 1/2)
}
\arguments{
\item{S}{information on the STOP region}

\item{U}{information on the UNDO region}

\item{X}{information on the x scale}

\item{Y}{information on the y scale}

\item{bell}{TRUE to ring the bell with mouse clicks}

\item{verbose}{TRUE to print more information}

\item{type}{'p' for points, 'n' for none, etc as for \code{\link{plot}}}

\item{col}{colour (ignored unless plotting) as for \code{\link{par}}}

\item{lwd}{line width of symbols as for \code{\link{par}}}

\item{pch}{symbol code as for \code{\link{par}}}

\item{cex}{symbol size expansion factor as 0for \code{\link{par}}}
}
\value{
a list containing digitized values \code{x} and \code{y} along with \code{status}, which may be \code{"DATA"} for data, \code{"AXIS"} for axis information, \code{"STOP"} for a click in the STOP region, or \code{"UNDO"} for a click in the UNDO region.
plus other properties
}
\description{
Digitize a point, or UNDO, or STOP.
}

