\name{RiceRaw.dat}
\docType{data}
\alias{RiceRaw.dat}
\title{Data for an experiment to investigate a rice germplasm panel}
\description{
The data is from an experiment in a Smarthouse in the Plant Accelerator. It is described in Al-Tamimi et al. (2016). It is used in \code{\link{imageData-package}} as an executable example to illustrate the use of \code{imageData}.

}
\usage{data(RiceRaw.dat)}
\format{A data.frame containing 14784 observations on 33 variables.}
\references{
Al-Tamimi, N, Brien, C.J., Oakey, H., Berger, B., Saade, S., Ho, Y. S., Schmockel, S. M., Tester, M. and Negrao, S. (2016) New salinity tolerance loci revealed in rice using high-throughput non-invasive phenotyping. \emph{Nature Communications}, \bold{7}, 13342.
}
\source{It will be made available on Dryad}
\keyword{datasets}