\name{topright}
\alias{topright}
\title{A helper function to specify the stop region.}
\usage{
topright()
}
\value{
a list with x and y page coordinates
}
\description{
topright invites the user to click the top-right corner of
the plotting area, and stores the x and y values (in plot
units) as xstop and ystop. These values are used by data().
}

