\name{bottomleft}
\alias{bottomleft}
\title{A helper function to specify the undo region.}
\usage{
bottomleft()
}
\value{
a list with x and y page coordinates
}
\description{
bottomleft invites the user to click the bottom-leftcorner
of the plotting area, and stores the x and y values (in
plot units) as xundo and yundo. These values are used by
data().
}

