% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getangle.R
\name{getangle}
\alias{getangle}
\title{Find the angle of a plot}
\usage{
getangle(image)
}
\arguments{
\item{image,}{the image to be worked with}
}
\value{
counterclockwise angle in degrees
}
\description{
'getangle' finds the counterclockwise angle at which the plot has
been rotated, determining the value from the user via pointer clicks.
}

