\name{getangle}
\alias{getangle}
\title{Find the angle of a plot}
\usage{
getangle(image, verbose = TRUE)
}
\arguments{
  \item{image,}{the image to be worked with}

  \item{verbose}{a logical indicating whether to indicate
  data as clicked}
}
\value{
counterclockwise angle in degrees
}
\description{
'getangle' finds the counterclockwise angle at which the
plot has been rotated, determining the value from the user
via pointer clicks.
}

